all: use go:embed for named.root file
