--- generate.sh
@@ -1,4 +1,0 @@
-#!/usr/bin/env sh
-echo "package dnsr\n\nvar root = \`" > root.go
-curl https://www.internic.net/domain/named.root >> root.go
-echo "\`" >> root.go
